Significantly improved the behavior of the script for handling infs and nans. These values are now interpolated during the reordering procedure and then casted back to the original data after each sweep.
The ebounds functionality is modified to ignore specific states outside the energy bounds at each point

The reordering metric uses a mean instead of a sum to prevent values from blowing up as quickly. Logarithms would be more effective, however the relationship to product rule of differentiation is less clear.